% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_microdata.R
\name{download_microdata}
\alias{download_microdata}
\title{Download Microdata}
\usage{
download_microdata(dir = tempdir(), overwrite = TRUE)
}
\arguments{
\item{dir}{The directory where the microdata will be downloaded.}

\item{overwrite}{Logical. If TRUE, the function will overwrite the file if it
already exists. If FALSE, the function will stop and return an error if the
file already exists.}
}
\value{
A tibble with the available microdata by country and year.
}
\description{
Checks the most recent available microdata uploaded to our
 GitHub repository.
}
